fixed bugs

+++ src/main/java/com/ood/Model/Currency/USD_Currency.java
@@ -13,7 +13,7 @@ public class USD_Currency extends AbsCurrency{
         this.amount-=valueOf(otherAmount,type);
     }
 
-    public static float valueOf(float otherAmount, CurrencyEnum type) {
+    public static double valueOf(double otherAmount, CurrencyEnum type) {
         if(type.equals(CurrencyEnum.JPY))
         {
             return otherAmount*Constants.EXCHANGE_RATE_JPY_DIV_USD;
